Renames components filenames
